backend/requirements.txt
flask>=3.0.0,<4.0.0
gunicorn>=22.0.0,<23.0.0
openai>=1.40.0,<2.0.0
python-dotenv>=1.0.1,<2.0.0
PyYAML>=6.0.0,<7.0.0
pytest>=8.0.0,<9.0.0
qdrant-client>=1.10.0,<2.0.0
semantic-kernel>=1.10.0,<2.0.0
asgiref>=3.8.1,<4.0.0
psycopg[binary]>=3.2.0,<4.0.0
psycopg-pool>=3.2.0,<4.0.0
werkzeug>=3.0.0,<4.0.0
pypdf>=5.4.0,<6.0.0
